LayoutTests/http/tests/websocket/tests/bad-sub-protocol-control-chars-expected.txt

CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket '\u0000'
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket '\u0009'
CONSOLE MESSAGE: line 0: Wrong protocol for WebSocket '\u001B'
Test WebSocket bad sub-protocol names by control characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "\0") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "	") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Parse error.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "
") threw exception SyntaxError: Parse error.
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS successfullyParsed is true

TEST COMPLETE
